Add Home Assistant as a community install option
Home Assistant is an open source home automation framework commonly installed on Raspberry Pi devices as an open source home automation controller.

It has an extensive Add-On package manager allowing people to install various extensions. Actual Budget's architexture makes it a good candidate as an add-on:

* Add-ons are docker containers, and actual can already run as a single docker container
* It only needs access to `/data` and a single http port to run - ideal condition for HA Add-Ons
* The server is actually only used for syncing, calculations and computational intensive features run on the client machines, so running it would not take away too much resources on limited machines, like the RPi

I have made an Add-On repository for Home Assistant that includes both the latest stable and latest edge version of Actual Budget that can be added to Home Assistant to be installed. The repository uses GitHub Actions to keep itself updated whenever new images are pushed to gchr.io.

Note that the add-on already works quite well without any extra steps, but because by default it only runs in unssecured http mode there is a nag screen that this is unsupported (and there is actually a need to occasionally reset the budget cache), so a README was also added to the root to show people how they can use other Home Assitant Add-Ons to run the server behind a properly verified Https proxy using Let's Encrypt

This change just adds a link to this README allowing people to follow the guides and install Actual on their HA instances