Clarified how often budgeteers can pull data via GoCardless
Previous documentation listed "up to 50 pulls a month". That is not the case. According to https://gocardless.com/bank-account-data/ and https://bankaccountdata.zendesk.com/hc/en-gb/articles/11528933493916-Bank-Account-Data-API-Usage-how-is-your-usage-number-calculated the limit is on connections, but the number of pulls is unlimited within those connections (nevermind that fact that the banks may rate limit the number of pulls, which they seem to do).