Update local.md - Adding Linux systemd unit file instructions
I wrote some steps on how to automate Actual starting up with Linux systems using systemd. I'm using this on my server and so far so good.

If this is too much for this page please consider putting it on its own page. I think it's useful info for anyone who wants to do self-hosted on a VPS with a non-Docker instance.