Update https.md
Add instructions for creating the config.json via a single terminal command. 

Reasoning - Assisted a MacOS deploy where they had everything setup correctly but when they made the config.json file their instance crashed, this was due to using the default GUI text editor which saved the file as a .rtf and they renamed to .json. Doing this will leave metadata in the file and actual wont know what to do with it causing an infinite restart loop. 

This addition provides a simple method to create the file without the possibility of extra formatting bits causing issues