Remove newline ambiguity from examples
The last two examples for the goals have created confusion
because the template and goals line shows in one line:
https://discord.com/channels/937901803608096828/1372691110614274170/1376211790639136779

This commit makes them display above each other, to make it clear that
they won't work when typed into one line.
It also does the same for two template examples
in the 'Notes on Priorities' section.